tracing_attributes

Module expand

Structs

AsyncInfo 🔒

IdentAndTypesRenamer 🔒

A visitor struct to replace idents and types in some piece of code (e.g. the “self” and “Self” tokens in user-supplied fields expressions when the function is generated by an old version of async-trait).

ImplTraitEraser 🔒

Enums

AsyncKind 🔒

The specific async code pattern that was detected

RecordType 🔒

Indicates whether a field should be recorded as Value or Debug.

Functions

erase_impl_trait 🔒

gen_block 🔒

Instrument a block

gen_function 🔒

Given an existing function, generate an instrumented version of that function

param_names 🔒

path_to_string 🔒