tracing_attributes

Function instrument_precise

pub(crate) fn instrument_precise(
    args: InstrumentArgs,
    item: TokenStream,
) -> Result<TokenStream, Error>

Instrument the function, by fully parsing the function body, which allows us to rewrite some statements related to async-like patterns.