Crate tracing_attributes

A procedural macro attribute for instrumenting functions with tracing.

tracing is a framework for instrumenting Rust programs to collect structured, event-based diagnostic information. This crate provides the #[instrument] procedural macro attribute.

Note that this macro is also re-exported by the main tracing crate.

Compiler support: requires rustc 1.56+

Usage

In the Cargo.toml:

[dependencies]
tracing-attributes = "0.1.24"

The #[instrument] attribute can now be added to a function to automatically create and enter tracing span when that function is called. For example:

use tracing::instrument;

#[instrument]
pub fn my_function(my_arg: usize) {
    // ...
}

Supported Rust Versions

Tracing is built against the latest stable release. The minimum supported version is 1.56. The current Tracing version is not guaranteed to build on Rust versions earlier than the minimum supported version.

Tracing follows the same compiler support policies as the rest of the Tokio project. The current stable Rust compiler and the three most recent minor versions before it will always be supported. For example, if the current stable compiler version is 1.69, the minimum supported version will not be increased past 1.66, three minor versions prior. Increasing the minimum supported compiler version is not considered a semver breaking change as long as doing so complies with this policy.

Attribute Macros

instrument

Instruments a function to create and enter a tracing span every time the function is called.